Question endpoint

+++ questions/views.py
@@ -1,5 +1,6 @@
 from django.conf import settings
 from django.contrib.auth import authenticate
+from django.db.models import Prefetch
 from rest_framework import status
 from rest_framework.decorators import api_view, authentication_classes, permission_classes
 from rest_framework.response import Response
@@ -59,45 +60,42 @@ def question_list(request):
         return Response(serializer.data, status=status.HTTP_200_OK)
     
     elif request.method == 'POST':
-        data = request.data
-        data['author'] = user.id
-        serializer = QuestionsSerializer(data=data)
+        serializer = QuestionsSerializer(data=request.data)
+        
         if serializer.is_valid():
-            serializer.save()
+            serializer.save(author=user)
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
-@api_view(['GET', 'PUT', 'DELETE'])
+@api_view(['GET', 'DELETE'])
 @permission_classes([IsAuthenticated])
 def question_detail(request, question_id):
     """
-    Retrieve, update or delete a question
+    Retrieve and delete a question
     """
     try: 
-        question = Questions.objects.get(pk=question_id)
+        question = Questions.objects.prefetch_related('answers_set').get(pk=question_id)
     except Questions.DoesNotExist:
         return Response(status=status.HTTP_404_NOT_FOUND)
     
     if request.method == 'GET':
         serializer = QuestionsSerializer(question)
-        return Response(serializer.data)
-                        
-    elif request.method == 'PUT':
-        serializer = QuestionsSerializer(question, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-    
+        return Response(serializer.data, status.HTTP_200_OK)
+     
     elif request.method == 'DELETE':
-        question.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if request.user.id == question.id: 
+            question.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        else:
+            return Response(status=status.HTTP_404_NOT_FOUND)
     
 
 @api_view(['GET', 'POST'])  
 @authentication_classes([])
 def answer_question(request, question_id):
+    """
+    """
     pass 
 
 @api_view(['GET', 'POST'])
